Rearranged EvaProperty to include all properies prefixed by "eva."
