+++ eva-server/src/test/resources/application.properties
@@ -1,10 +1,10 @@
 eva.mongo.host=testHost
 eva.mongo.user=testUser
 eva.mongo.passwd=testPasswd
-#eva.mongo.auth.db=@eva.mongo.auth.db@
+eva.mongo.auth.db=testAuthDb
 eva.mongo.read-preference=testReadPreference
 
-#eva.mongo.collections.variants=testCollections
-#eva.mongo.collections.files=@eva.mongo.collections.files@
-#
-#eva.version=@eva.version@
+eva.mongo.collections.variants=testVariants
+eva.mongo.collections.files=testFiles
+
+eva.version=testVersion
